# [WIP] Small memory footprint GC optimization level