--- a/src/gc/gc.cpp
+++ b/src/gc/gc.cpp
@@ -3092,61 +3092,6 @@ gc_heap::dt_low_ephemeral_space_p (gc_tuning_point tp)
     return ret;
 }
 
-float
-gc_heap::dt_gen2_frag_tolerance(gc_tuning_point tp)
-{
-    if (tp == tuning_deciding_condemned_gen)
-    {
-        switch (latency_level)
-        {
-            // if we are tuning for short pauses, we are much
-            // more tolerant of gen2 frag.
-            case latency_level_short_pauses:
-                return 0.80f;
-            case latency_level_balanced:
-            case latency_level_throughput:
-                return 0.65f;
-            // if we are tuning for low memory footprint, we
-            // are much /less/ tolerant of gen2 frag.
-            case latency_level_memory_footprint:
-                return 0.40f;
-            default:
-                assert(!"invalid value for latency_level");
-                return 0.65f;
-        }
-    }
-
-    assert(!"asking for frag tolerance at unexpected tuning point");
-    return 0;
-}
-
-BOOL
-gc_heap::dt_gen2_high_frag_p(gc_tuning_point tp)
-{
-    BOOL ret = FALSE;
-
-    switch (tp)
-    {
-        case tuning_deciding_condemned_gen:
-        {
-            float frag_ratio = (float)dd_fragmentation (dynamic_data_of (max_generation)) / (float)generation_size (max_generation);
-            float frag_tolerance = dt_gen2_frag_tolerance (tp);
-            if (frag_ratio > frag_tolerance)
-            {
-                dprintf (GTC_LOG, ("g2 FR: %d%%", (int)(frag_ratio*100)));
-                dprintf (GTC_LOG, ("g2 frag tolerance: %d%%", (int)(frag_tolerance*100)));
-                return TRUE;
-            }
-
-            return FALSE;
-        }
-        default:
-            break;
-    }
-
-    return ret;
-}
-
 BOOL 
 gc_heap::dt_high_frag_p (gc_tuning_point tp, 
                          int gen_number, 
@@ -3170,9 +3115,14 @@ gc_heap::dt_high_frag_p (gc_tuning_point tp,
             else
             {
 #ifndef MULTIPLE_HEAPS
-                if (gen_number == max_generation && dt_gen2_high_frag_p (tp))
+                if (gen_number == max_generation)
                 {
-                    return TRUE;
+                    float frag_ratio = (float)(dd_fragmentation (dynamic_data_of (max_generation))) / (float)generation_size (max_generation);
+                    if (frag_ratio > 0.65)
+                    {
+                        dprintf (GTC_LOG, ("g2 FR: %d%%", (int)(frag_ratio*100)));
+                        return TRUE;
+                    }
                 }
 #endif //!MULTIPLE_HEAPS
                 size_t fr = generation_unusable_fragmentation (generation_of (gen_number));
@@ -10166,7 +10116,7 @@ gc_heap::init_semi_shared()
     should_expand_in_full_gc = FALSE;
 
 #ifdef FEATURE_LOH_COMPACTION
-    loh_compaction_always_p = GCConfig::GetLOHCompactionMode() != 0;
+    loh_compaction_always_p = GCConfig::GetLOHCompactionMode() != 0 || gc_heap::latency_level == latency_level_small_memory_footprint;
     loh_compaction_mode = loh_compaction_default;
 #endif //FEATURE_LOH_COMPACTION
 
@@ -30326,7 +30276,7 @@ BOOL gc_heap::decide_on_compacting (int condemned_gen_number,
 #endif // GC_STATS
 #endif //STRESS_HEAP
 
-    if (GCConfig::GetForceCompact())
+    if (GCConfig::GetForceCompact() || gc_heap::latency_level == latency_level_small_memory_footprint)
         should_compact = TRUE;
 
     if ((condemned_gen_number == max_generation) && last_gc_before_oom)
@@ -33524,6 +33474,13 @@ HRESULT GCHeap::Initialize ()
 {
     HRESULT hr = S_OK;
 
+    // Get our latency level. Defaults to latency_level_balanced if not specified or if the config latency level is invalid.
+    int latency_level_from_config = static_cast<int>(GCConfig::GetLatencyLevel());
+    if (latency_level_from_config >= latency_level_first && latency_level_from_config <= latency_level_last)
+    {
+        gc_heap::latency_level = static_cast<gc_latency_level>(latency_level_from_config);
+    }
+
     g_gc_pFreeObjectMethodTable = GCToEEInterface::GetFreeObjectMethodTable();
     g_num_processors = GCToOSInterface::GetTotalProcessorCount();
     assert(g_num_processors != 0);
@@ -33586,15 +33543,6 @@ HRESULT GCHeap::Initialize ()
         available_mem_th = min (available_mem_th, adjusted_available_mem_th);
     }
 
-    // Get our latency level. Defaults to 3 if not specified, and falls-back to 3
-    // if the config latency level is invalid.
-    int latency_level_from_config = static_cast<int>(GCConfig::GetLatencyLevel());
-    if (latency_level_from_config < 1 || latency_level_from_config > 4)
-    {
-        latency_level_from_config = 3;
-    }
-
-    gc_heap::latency_level = static_cast<gc_latency_level>(latency_level_from_config);
     gc_heap::high_memory_load_th = 100 - available_mem_th;
 
 #if defined(BIT64) 
@@ -35056,9 +35004,9 @@ GCHeap::GarbageCollectGeneration (unsigned int gen, gc_reason reason)
 
     gc_heap::g_low_memory_status = (reason == reason_lowmemory) || 
                                    (reason == reason_lowmemory_blocking) ||
-                                   g_bLowMemoryFromHost;
+                                   gc_heap::latency_level == latency_level_small_memory_footprint;
 
-    if (g_bLowMemoryFromHost)
+    if (gc_heap::latency_level == latency_level_small_memory_footprint)
         reason = reason_lowmemory_host;
 
     gc_trigger_reason = reason;