--- a/src/gc/gcpriv.h
+++ b/src/gc/gcpriv.h
@@ -422,6 +422,14 @@ enum gc_type
     gc_type_max = 3
 };
 
+enum gc_latency_level
+{
+    latency_level_memory_footprint = 1,
+    latency_level_throughput       = 2,
+    latency_level_balanced         = 3,
+    latency_level_short_pauses     = 4
+};
+
 #define v_high_memory_load_th 97
 
 //encapsulates the mechanism for the current gc
@@ -3419,6 +3427,12 @@ protected:
     BOOL dt_low_card_table_efficiency_p (gc_tuning_point tp);
 
     PER_HEAP
+    BOOL dt_gen2_high_frag_p(gc_tuning_point tp);
+
+    PER_HEAP_ISOLATED
+    float dt_gen2_frag_tolerance(gc_tuning_point tp);
+
+    PER_HEAP
     int generation_skip_ratio;//in %
 
     PER_HEAP
@@ -3551,6 +3565,24 @@ protected:
     // TODO: get rid of total_ephemeral_plugs.
     PER_HEAP
     size_t total_ephemeral_size;
+
+    // The latency level requested by user. Levels are numbered "1" through "4"
+    // and have the following meanings:
+    // +----------+--------------------+---------------------------------------+
+    // | Level    | Optimization Goals | Latency Charactaristics               |
+    // +----------+--------------------+---------------------------------------+
+    // | 1        | memory footprint   | pauses can be long and more frequent  |
+    // | 2        | throughput         | pauses are unpredictable and not very |
+    // |          |                    | frequent, and might be long           |
+    // | 3        | balanced           | pauses are more predictable and more  |
+    // |          |                    | frequent. the longest pauses are      |
+    // |          |                    | shorter than 2.                       |
+    // | 4        | short pauses       | pauses are more predictable and more  |
+    // |          |                    | frequent. the longest pauses are      |
+    // |          |                    | shorter than 3.                       |
+    // +----------+--------------------+---------------------------------------+
+    PER_HEAP_ISOLATED
+    gc_latency_level latency_level;
 
 public: