This pull request adds small memory footprint as GC optimization level.

This pull request is based on the branch, mentioned by @swgillespie in https://github.com/dotnet/coreclr/issues/13292#issuecomment-324108583, but part with fragmentation tuning was removed because its combination with compacting was not researched.

@Maoni0 @swgillespie, please, take a look