@swgillespie 2017-08-29
I'll defer to @Maoni0 on this one. Do you think we're ready to take this or should we try and exercise the `latency_level_small_memory_footprint_unstable_perf` first?