@gbalykov 2017-09-04
@dotnet-bot test CentOS7.1 x64 Debug Build and Test
@dotnet-bot test Ubuntu x64 Checked Build and Test