@Maoni0 2017-09-15
@gbalykov, @swgillespie, I pushed a [commit in my branch](https://github.com/Maoni0/coreclr/commit/03bfe764a5f67fa671c0c7092d47a1624005cbb9) that shows how the latency levels should be set up. you can see lots of TODO's in the code. I incorporated some of @gbalykov's changes (unfortunately I dunno how to get both your commits and make them show up as from you guys :P and I am running out of time before I go on vacation; but there weren't many changes I incorporated so it should be fine if one of you wants to take this and keep working on it). 

As you can see, I've separated out the static data for the latency levels - for now I've kept everything the same as before except the data for the memory_footprint level since we already know some of what we want to change - I expect us to be doing lots of perf tuning to set up the appropriate static data for each level (I also changed the values for the enums... I am actually thinking instead of making these at increments of 1 perhaps we should make them increments of 10 to give us more numbers for each level so we can have more detailed levels for optimizing for memory footprint and etc). 

I intentionally did not take the change that always does compacting for this level because we don't want to "always compact" - there are situations where we know compacting is completely uselss so it's not a good time to unconditionally compact; instead I changed the fragment static data to be smaller so it would compact more often.