@gbalykov 2017-09-27
@Maoni0 @swgillespie

I have measured the branch on few Tizen GUI applications, and results are the next

Application|Original, Peak, Private_Dirty, Kb|Modified (GCLatencyLevel=memory_footprint), Peak, Private_Dirty, Kb|Difference
-----------|---------------------------------|---------------------------------|----------
HelloWorld|396|392|4 (1.01%)
BoxViewClock|1404|760|644 (45.87%)
Puzzle|912|600|312 (34.21%)
ApplicationStoreUI|524|492|32 (6.11%)
GalleryUI|424|392|32 (7.55%)
AppCommon|764|616|148 (19.37%)
SNSUI|2364|1632|732 (30.96%)
Calculator|1080|712|368 (34.07%)

average: 22.39%

I think this results are good to start with and I will continue to work on this after finishing another CoreCLR optimization.