@gbalykov 2017-10-09
@Maoni0 @swgillespie

I have also checked on the balanced level and it seems to increase memory consumption. 

Application|Original, Peak, Private_Dirty, Kb|Modified (GCLatencyLevel=balanced), Peak, Private_Dirty, Kb|Difference
-----------|---------------------------------|---------------------------------|----------
HelloWorld|396|448| -52 (-13.13%)
BoxViewClock|1404|2252| -847 (-60.33%)
Puzzle|912|1136| -224 (-24.56%)
ApplicationStoreUI|524|696| -172 (-32.82%)
GalleryUI|424|600| -176 (-41.51%)
AppCommon|764|944| -180 (-23.56%)
SNSUI|2364|2512| -148 (-6.26%)
Calculator|1080|1264| -184 (-17.04%)

average memory consumption increase: 27.40%