@Maoni0 2017-10-09
Yes, that's expected for memory consumption. Do you also have throughput numbers?