@gbalykov 2017-11-28
@Maoni0 

Sorry for the delay, I've been busy doing another optimization.

These are the results for your patch (base 97d9d593ae44438fab85044e2a189316e7f3cbc4):

Test|base GC heap (Kb), startup time (s)|"memory_consumption" level GC heap (Kb), startup time (s)
----|---------|----
HelloWorld.Tizen|368; 1.99158|288 (-21.74%); 1.9505 (-2.06%)
Settings.Tizen.Mobile|632; 2.54684|520 (-17.72%); 2.522 (-0.98%)
VolumeController.Tizen.Mobile|512; 2.391|472 (-7.81%); 2.343 (-2.01%)
AppCommon.Tizen.Mobile|838.4; 3.824|612 (-27.00%); 3.8325 (+0.22%)
EmailUI.Tizen.Mobile|1384; 3.1115|1016 (-26.59%); 3.0865 (-0.80%)
ApplicationControl.Tizen.Mobile|706.8; 3.2635|552 (-21.90%); 3.229 (-1.06%)
SNSUI.Tizen.Mobile|2292; 4.424|1732.4 (-24.42%); 4.511 (+1.97%)
ApplicationStoreUI.Tizen.Mobile|1203.2; 2.425|884 (-26.53%); 2.41842 (-0.27%)
Puzzle.Tizen.Mobile|970; 3.001|701.2 (-27.71%); 3.0 (-0.03%)
System_info.Tizen.Mobile|368; 2.366|288 (-21.74%); 2.3355 (-1.29%)
GalleryUI.Tizen.Mobile|368; 2.4585|328 (-10.87%); 2.3985 (-2.44%)
Calculator.Tizen.Mobile|2078.6; 4.855|792 (-61.90%); 5.1805 (+6.70%)

average: -24.66%; -0.17%

Test|base GC heap (Kb), startup time (s)|"balanced" level GC heap (Kb), startup time (s)
----|---------|----
HelloWorld.Tizen|368; 1.99158|388 (+5.43%); 1.93 (-3.09%)
Settings.Tizen.Mobile|632; 2.54684|816 (+29.11%); 2.4945 (-2.06%)
VolumeController.Tizen.Mobile|512; 2.391|696 (+35.94%); 2.32526 (-2.75%)
AppCommon.Tizen.Mobile|838.4; 3.824|1024 (+22.14%); 3.802 (-0.58%)
EmailUI.Tizen.Mobile|1384; 3.1115|1560 (+12.72%); 3.0655 (-1.48%)
ApplicationControl.Tizen.Mobile|706.8; 3.2635|890.4 (+25.98%); 3.228 (-1.09%)
SNSUI.Tizen.Mobile|2292; 4.424|2537 (+10.69%); 4.4175 (-0.15%)
ApplicationStoreUI.Tizen.Mobile|1203.2; 2.425|1388.4 (+15.39%); 2.39053 (-1.42%)
Puzzle.Tizen.Mobile|970; 3.001|1155.6 (+19.13%); 2.968 (-1.10%)
System_info.Tizen.Mobile|368; 2.366|444 (+20.65%); 2.329 (-1.56%)
GalleryUI.Tizen.Mobile|368; 2.4585|540 (+46.73%); 2.4085 (-2.03%)
Calculator.Tizen.Mobile|2078.6; 4.855|5324 (+156.13%); 4.8145 (-0.83%)

average: +33.34%; -1.51%

</br>

Also, I have been performing tuning of parameters from static_data structure on Tizen Xamarin GUI applications and GC benchmarks and I always seem to obtain the same results with +-3% accuracy. I set fragmentation and fragmentation_burden for both gen0 and gen1 (for latency_level_memory_footprint) lower than the ones that you have specified (i.e. lower than 40000, 0.25, 80000, 0.25).

What could be the reason of this? Could it be due to reaching some kind of a minimum at your parameters, which couldn't be lowered any more?

Could you, please, also suggest a better way of tuning of parameters from static_data?

cc @kvochko